@@ -18,13 +18,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   dart_ipify: ^1.1.1
   device_info_plus: ^3.0.1
+  shared_preferences: ^2.0.7
+  contacts_service: ^0.6.3
+  permission_handler: ^8.1.5
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
